requirements.txt
sendgrid==6.12.5
python-dotenv==1.0.0
httpx==0.27.0
aiohttp==3.11.0
pydantic==2.10.5
gradio==4.44.0
